Vedant/ai_interview_tool/backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
redis==6.4.0
python-socketio==5.11.0
aiohttp==3.9.5
websockets==12.0
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.51.0
pyyaml==6.0.3
